build: Moving the strict DNS cluster into an extension directory
