+++ CODEOWNERS
@@ -278,6 +278,7 @@ extensions/filters/http/oauth2 @derekargueta @snowp
 /*/extensions/filters/common @UNOWNED @UNOWNED
 /*/extensions/filters/http/common @UNOWNED @UNOWNED
 /*/extensions/filters/network/common @UNOWNED @UNOWNED
+/*/extensions/clusters/strict_dns @UNOWNED @UNOWNED
 /*/extensions/clusters/original_dst @UNOWNED @UNOWNED
