+++ docs/root/intro/arch_overview/upstream/service_discovery.rst
@@ -20,6 +20,7 @@ Static is the simplest service discovery type. The configuration explicitly spec
 network name (IP address/port, unix domain socket, etc.) of each upstream host.
 
 .. _arch_overview_service_discovery_types_strict_dns:
+.. _extension_envoy.clusters.strict_dns:
 
 Strict DNS
 ^^^^^^^^^^
